+++ package/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - add JSDoc default options
 
+### What's Changed
+
+- better DX for importing constants
+
 ## 0.1.0 (2026-05-29)
 
 initial release